Moved agentset to library folder
